+++ scripts/bundles/sys-util.js
@@ -6,4 +6,6 @@ exports.isGlob = require('is-glob');
 
 exports.minimatch = require('minimatch');
 
+exports.minimist = require('minimist');
+
 exports.semver = require('semver');