+++ src/compiler/config/test/validate-asset-versioning.spec.ts
@@ -1,0 +1,28 @@
+import { Config } from '../../../declarations';
+import { mockLogger, mockStencilSystem } from '../../../testing/mocks';
+import { validateBuildConfig } from '../validate-config';
+import * as path from 'path';
+
+
+describe('validateAssetVerioning', () => {
+
+  let config: Config;
+  const logger = mockLogger();
+  const sys = mockStencilSystem();
+
+  beforeEach(() => {
+    config = {
+      sys: sys,
+      logger: logger,
+      rootDir: '/User/some/path/',
+      suppressTypeScriptErrors: true
+    };
+  });
+
+
+  it('should default null', () => {
+    validateBuildConfig(config);
+    expect(config.assetVersioning).toBe(null);
+  });
+
+});