+++ src/declarations/config.ts
@@ -16,6 +16,7 @@ export type ConfigBuildTarget = {
 };
 
 export interface Config {
+  assetVersioning?: ConfigAssetVersioning;
   buildAppCore?: boolean;
   buildEs5?: boolean;
   buildLogFilePath?: string;
@@ -77,6 +78,59 @@ export interface Config {
    * DEPRECATED "config.collections" since 0.6.0, 2018-02-13
    */
   _deprecatedCollections?: ConfigCollection[];
+}
+
+
+export interface ConfigAssetVersioning {
+  /**
+   * An array of strings of which CSS properties to version.
+   */
+  cssProperties?: string[];
+
+  /**
+   * The length of the hash. Defaults to config.hashedFileNameLength.
+   */
+  hashLength?: number;
+
+  /**
+   * If true, the plugin will put the hash to the query string instead of the filename. Defaults to false;
+   */
+  queryMode?: boolean;
+
+  /**
+   * Glob search pattern for urls to version. Defaults to include css,js,png,jpg,jpeg,gif,svg,json,woff,woff2,eot,ttf
+   */
+  pattern?: string;
+
+  /**
+   * Function used to filter which assets to version. First argument is the url. Return true for the url to be versioned. The "pattern" config will not be used if a "filter" function is given.
+   */
+  filter?: (url: string) => boolean;
+
+  /**
+   * Function used to create the filename with the file's hash. First argument is the filename, and the second argument is the hash.
+   */
+  generateFileName?: (fileName: string, hash: string) => string;
+
+  /**
+   * The prefix to prepended to the file path. Defaults to "".
+   */
+  prefix?: string;
+
+  /**
+   * The separator between the filename and hash. Defaults to "."
+   */
+  separator?: string;
+
+  /**
+   * Version assets found in HTML elements, such as CSS urls in <link> and JS urls in <script>. Defaults to true;
+   */
+  versionHtml?: boolean;
+
+  /**
+   * Version assets that are found in styles and css, such as background-url and @font-face urls. Defaults to true;
+   */
+  versionCssProperties?: boolean;
 }